+++ packages/orq-rc/src/orq_ai_sdk/models/fileuploadop.py
@@ -94,5 +94,5 @@ class FileUploadResponseBody(BaseModel):
     workspace_id: str
     r"""The id of the resource"""
 
-    created: Optional[datetime] = dateutil.parser.isoparse("2025-03-13T18:09:36.739Z")
+    created: Optional[datetime] = dateutil.parser.isoparse("2025-03-13T22:21:46.762Z")
     r"""The date and time the resource was created"""