+++ packages/orq-rc/RELEASES.md
@@ -849,3 +849,13 @@ Based on:
 - [python v3.2.0-rc.47] packages/orq-rc
 ### Releases
 - [PyPI v3.2.0-rc.47] https://pypi.org/project/orq-ai-sdk/3.2.0-rc.47 - packages/orq-rc
+
+## 2025-03-13 22:22:57
+### Changes
+Based on:
+- OpenAPI Doc  
+- Speakeasy CLI 1.517.3 (2.548.6) https://github.com/speakeasy-api/speakeasy
+### Generated
+- [python v3.2.0-rc.48] packages/orq-rc
+### Releases
+- [PyPI v3.2.0-rc.48] https://pypi.org/project/orq-ai-sdk/3.2.0-rc.48 - packages/orq-rc